Remove unreachable reference to secure_headers gem (#7157)
* Remove unreachable reference to secure_headers gem

changelog: Internal, Code Quality, Remove unreachable code

* Re-expand map block

Avoid nested inline blocks, simplify diff
